Extract address in module
